Fixed issue with ad hoc files localization. Also performed some refactoring in order to reduce code duplication.
